-- Fixed compilation problem on demo drivers
   demo_drivers/interaction/pseudo_solid_collapsible_tubes
   demo_drivers/mpi/multi_domain/pseudo_solid_collapsible_tubes

-- Added validation data for triangle mesh internal boundaries.

   demo_drivers/meshing/mesh_from_inline_triangle_internal_boundaries

-- Fixed bugs when establishing the refinement tolerance on
   TriangleMeshCurve

   - TriangleMeshCurve::enable_polyline_refinement_tolerance()
   - TriangleMeshCurve::enable_polyline_unrefinement_tolerance()
   - TriangleMeshCurve::set_polyline_refinement_tolerance()
   - TriangleMeshCurve::set_polyline_unrefinement_tolerance()
   - TriangleMeshCurve::disable_polyline_refinement()
   - TriangleMeshCurve::disable_polyline_unrefinement()
   
   


git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@299 595e1d82-adf1-4282-bac7-b6073fa2ca8e

+++ src/meshes/triangle_mesh.template.h
@@ -1429,8 +1429,8 @@ class TriangleMeshParameters
        for (unsigned h = 0; r_it!=zeta_values_pt.end(); h++)
         {
          local_n_vertices =
-           ((double)n_t_vertices * delta_z[h]) /
-           std::fabs(zeta_final - zeta_initial);
+          static_cast<unsigned>(((double)n_t_vertices * delta_z[h]) /
+                    std::fabs(zeta_final - zeta_initial));
 
          local_zeta_initial = *l_it;
          local_zeta_final = *r_it;
@@ -1580,7 +1580,8 @@ class TriangleMeshParameters
           }
 
          local_n_vertices =
-           ((double)n_t_vertices * delta_z[h]) / (total_arclength);
+          static_cast<unsigned>(((double)n_t_vertices * delta_z[h]) / 
+                                (total_arclength));
 
          // Desired arclength increment
          double local_target_s_increment=
